+++ functions.php
@@ -121,8 +121,10 @@ function whiteboard_scripts() {
 
 	wp_enqueue_script( 'whiteboard-skip-link-focus-fix', get_template_directory_uri() . '/js/skip-link-focus-fix.js', array(), '20130115', true );
 
-  // bower:js
-  // endbower
+  // scripts
+  wp_enqueue_script( 'whiteboard-jquery', get_template_directory_uri() . '/js/vendors/jquery/jquery.js', array(), '20130115', true );
+  wp_enqueue_script( 'whiteboard-lodash', get_template_directory_uri() . '/js/vendors/lodash/lodash.js', array(), '20130115', true );
+  // endscripts
 
 	if ( is_singular() && comments_open() && get_option( 'thread_comments' ) ) {
 		wp_enqueue_script( 'comment-reply' );